test(cli): separate login and logout cmd tests into cmd_test.go
