+++ internal/occ/cmd/login/login_test.go
@@ -91,52 +91,6 @@ func TestNewAuthImpl(t *testing.T) {
 	assert.NotNil(t, impl)
 }
 
-// --- NewLoginCmd ---
-
-func TestNewLoginCmd_Structure(t *testing.T) {
-	cmd := NewLoginCmd()
-
-	assert.Equal(t, "login", cmd.Use)
-	assert.NotEmpty(t, cmd.Short)
-	assert.NotNil(t, cmd.RunE)
-}
-
-func TestNewLoginCmd_Flags(t *testing.T) {
-	cmd := NewLoginCmd()
-
-	for _, name := range []string{"client-credentials", "client-id", "client-secret", "credential"} {
-		assert.NotNil(t, cmd.Flags().Lookup(name), "expected flag %q", name)
-	}
-}
-
-func TestNewLoginCmd_RunE_NoConfig(t *testing.T) {
-	testutil.SetupTestHome(t)
-
-	cmd := NewLoginCmd()
-	err := cmd.RunE(cmd, nil)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "failed to get current context")
-}
-
-func TestNewLoginCmd_RunE_ClientCredentials_MissingCreds(t *testing.T) {
-	home := testutil.SetupTestHome(t)
-	t.Setenv("OCC_CLIENT_ID", "")
-	t.Setenv("OCC_CLIENT_SECRET", "")
-
-	testutil.WriteOCConfig(t, home, &config.StoredConfig{
-		CurrentContext: "ctx",
-		ControlPlanes:  []config.ControlPlane{{Name: "cp", URL: "http://localhost"}},
-		Credentials:    []config.Credential{{Name: "cred"}},
-		Contexts:       []config.Context{{Name: "ctx", ControlPlane: "cp", Credentials: "cred"}},
-	})
-
-	cmd := NewLoginCmd()
-	require.NoError(t, cmd.Flags().Set("client-credentials", "true"))
-	err := cmd.RunE(cmd, nil)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "client ID and client secret are required")
-}
-
 func TestGetLoginPrompt(t *testing.T) {
 	impl := NewAuthImpl()
 	prompt := impl.GetLoginPrompt()
